move pseudo to advanced setting
