+++ src/aiidalab_qe/app/main.py
@@ -90,14 +90,6 @@ class App(ipw.VBox):
             (self.configure_qe_app_work_chain_step, "advanced_settings"),
             (self.submit_qe_app_work_chain_step, "advanced_settings"),
         )
-        ipw.dlink(
-            (self.configure_qe_app_work_chain_step, "pseudo_family_selector"),
-            (self.submit_qe_app_work_chain_step, "pseudo_family_selector"),
-        )
-        ipw.dlink(
-            (self.configure_qe_app_work_chain_step, "pseudo_setter"),
-            (self.submit_qe_app_work_chain_step, "pseudo_setter"),
-        )
 
         ipw.dlink(
             (self.submit_qe_app_work_chain_step, "process"),
